Split module and cli into separate files
